refact: move Scaffold and related structures to platform.theme.ui.scaffold
