+++ feature/composer/src/main/java/com/jeanbarrossilva/orca/feature/composer/Composer.kt
@@ -50,10 +50,10 @@ import com.jeanbarrossilva.orca.feature.composer.ui.Toolbar
 import com.jeanbarrossilva.orca.platform.theme.OrcaTheme
 import com.jeanbarrossilva.orca.platform.theme.extensions.backwardsNavigationArrow
 import com.jeanbarrossilva.orca.platform.theme.extensions.plus
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.Scaffold
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.bar.top.TopAppBar
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.bar.top.text.AutoSizeText
 import com.jeanbarrossilva.orca.platform.ui.component.input.TextFieldDefaults as _TextFieldDefaults
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.Scaffold
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.bar.TopAppBar
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.bar.text.AutoSizeText
 import com.jeanbarrossilva.orca.platform.ui.core.requestFocusWithDelay
 import com.jeanbarrossilva.orca.platform.ui.html.HtmlAnnotatedString
 
@@ -105,15 +105,17 @@ private fun Composer(
         modifier,
         topAppBar = {
             @OptIn(ExperimentalMaterial3Api::class)
-            TopAppBar(
-                navigationIcon = {
-                    IconButton(onClick = onBackwardsNavigation) {
-                        Icon(OrcaTheme.Icons.backwardsNavigationArrow, contentDescription = "Back")
-                    }
-                }
-            ) {
-                AutoSizeText("Compose")
+            (
+                TopAppBar(
+        navigationIcon = {
+            IconButton(onClick = onBackwardsNavigation) {
+                Icon(OrcaTheme.Icons.backwardsNavigationArrow, contentDescription = "Back")
             }
+        }
+    ) {
+        AutoSizeText("Compose")
+    }
+            )
         },
         floatingActionButton = {
             FloatingActionButton(
