+++ feature/toot-details/src/main/java/com/jeanbarrossilva/orca/feature/tootdetails/TootDetails.kt
@@ -27,10 +27,10 @@ import com.jeanbarrossilva.orca.platform.theme.extensions.backwardsNavigationArr
 import com.jeanbarrossilva.orca.platform.theme.reactivity.BottomAreaAvailabilityNestedScrollConnection
 import com.jeanbarrossilva.orca.platform.theme.reactivity.OnBottomAreaAvailabilityChangeListener
 import com.jeanbarrossilva.orca.platform.theme.reactivity.rememberBottomAreaAvailabilityNestedScrollConnection
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.bar.top.TopAppBar
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.bar.top.TopAppBarDefaults
+import com.jeanbarrossilva.orca.platform.theme.ui.scaffold.bar.top.text.AutoSizeText
 import com.jeanbarrossilva.orca.platform.ui.AccountFormatter
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.bar.TopAppBar
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.bar.TopAppBarDefaults
-import com.jeanbarrossilva.orca.platform.ui.component.scaffold.bar.text.AutoSizeText
 import com.jeanbarrossilva.orca.platform.ui.component.timeline.Timeline
 import com.jeanbarrossilva.orca.platform.ui.component.timeline.toot.TootPreview
 import com.jeanbarrossilva.orca.platform.ui.component.timeline.toot.formatted
@@ -140,19 +140,21 @@ private fun TootDetails(
         modifier,
         topBar = {
             @OptIn(ExperimentalMaterial3Api::class)
-            TopAppBar(
-                navigationIcon = {
-                    IconButton(onClick = onBackwardsNavigation) {
-                        Icon(
-                            OrcaTheme.Icons.backwardsNavigationArrow,
-                            contentDescription = "Back"
-                        )
-                    }
-                },
-                scrollBehavior = topAppBarScrollBehavior
-            ) {
-                AutoSizeText("Toot")
+            (
+                TopAppBar(
+        navigationIcon = {
+            IconButton(onClick = onBackwardsNavigation) {
+                Icon(
+                    OrcaTheme.Icons.backwardsNavigationArrow,
+                    contentDescription = "Back"
+                )
             }
+        },
+        scrollBehavior = topAppBarScrollBehavior
+    ) {
+        AutoSizeText("Toot")
+    }
+            )
         }
     ) {
         Timeline(
